refactor bot and ignore the error without db
